Added Birthday, Hannukah, Halloween, Christmas cards on components folder
