+++ client/src/pages/Members/index.js
@@ -1,8 +1,18 @@
 import React from "react";
-import Card from "../../components/Card";
+import Birthdaycard from '../../components/Birthday';
+import Halloweencard from '../../components/Halloween';
+import Hannukahcard from '../../components/Hannukah';
+import Christmascard from '../../components/Christmas';
 
 function Members() {
-  return <Card />;
+  return (
+    <div>
+      <Birthdaycard />
+      <Halloweencard />
+      <Hannukahcard />
+      <Christmascard />
+    </div>
+  )
 }
 
 export default Members;
